requirements.txt
fastmcp
httpx
python-dotenv
pydantic

